refactor: move metasrv options to common_options
