+++ src/meta-srv/Cargo.toml
@@ -20,6 +20,7 @@ common-error = { path = "../common/error" }
 common-grpc = { path = "../common/grpc" }
 common-grpc-expr = { path = "../common/grpc-expr" }
 common-meta = { path = "../common/meta" }
+common-options = { path = "../common/options" }
 common-procedure = { path = "../common/procedure" }
 common-runtime = { path = "../common/runtime" }
 common-telemetry = { path = "../common/telemetry" }
