+++ tests-integration/src/cluster.rs
@@ -23,6 +23,7 @@ use common_base::Plugins;
 use common_grpc::channel_manager::ChannelManager;
 use common_meta::peer::Peer;
 use common_meta::DatanodeId;
+use common_options::meta::MetaSrvOptions;
 use common_runtime::Builder as RuntimeBuilder;
 use common_test_util::temp_dir::create_temp_dir;
 use datanode::datanode::{DatanodeOptions, ObjectStoreConfig, ProcedureConfig};
@@ -33,7 +34,7 @@ use frontend::instance::{FrontendInstance, Instance as FeInstance};
 use meta_client::client::MetaClientBuilder;
 use meta_srv::cluster::MetaPeerClientRef;
 use meta_srv::metadata_service::{DefaultMetadataService, MetadataService};
-use meta_srv::metasrv::{MetaSrv, MetaSrvOptions};
+use meta_srv::metasrv::MetaSrv;
 use meta_srv::mocks::MockInfo;
 use meta_srv::service::store::kv::KvStoreRef;
 use meta_srv::service::store::memory::MemStore;
